requirements.txt
markdown2

fastapi
uvicorn
faiss-cpu
numpy
tiktoken
openai
python-dotenv
pillow
PyMuPDF
python-docx
google-cloud-vision
regex
python-multipart

SQLAlchemy
psycopg2-binary
pgvector
requests
beautifulsoup4

PyJWT
